User validation and exceptions

+++ src/main/java/ar/edu/unq/DevApp/service/CryptoService.java
@@ -1,6 +1,7 @@
 package ar.edu.unq.DevApp.service;
 
 import ar.edu.unq.DevApp.model.User;
+import ar.edu.unq.DevApp.model.exceptions.*;
 import ar.edu.unq.DevApp.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -15,11 +16,17 @@ public class CryptoService {
     UserRepository userRepository;
 
     public User createUser(User user) {
+        try {
+            user.validateUser();
+        } catch (InvalidNameException | InvalidEmailException | InvalidAddressException  | InvalidCVUException | InvalidPasswordException | InvalidWalletException e) {
+            throw e;
+        }
         return userRepository.save(user);
     }
 
-    public Optional<User> getUserById(Long id) {
-        return userRepository.findById(id);
+    public User getUserById(Long id) {
+        return userRepository.findById(id)
+                .orElseThrow(() -> new UserNotFoundException("User with ID " + id + " not found"));
     }
 
     public List<User> getAllUsers() {
@@ -27,6 +34,9 @@ public class CryptoService {
     }
 
     public void deleteUser(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new UserNotFoundException("User with ID " + id + " not found");
+        }
         userRepository.deleteById(id);
     }
